**refactor(isobmff): normalize box/length math to `int` and align readers/tests**
- convert ISOBMFF size/remain/length handling from mixed `int64`/`uint*` to `int` across box parsing, metadata traversal, CRX, PRVW, and XPacket paths
- update `XPacketHeader.Length` and related call sites/logging/tests to use `int`
- add explicit bounded conversions (`uint64ToInt`) and keep overflow checks centralized
- refresh ISOBMFF tests for new max-int behavior and remove redundant loop-var copies
- add targeted `nolint:gosec` annotations where casts are spec-bounded

**chore(lint): fix root benchmark/reader lint findings**

- add checked pooled-reader helper in `imagemeta.go` to satisfy `errcheck` type-assertion rules
- replace direct error equality checks with `errors.Is` in benchmarks (`errorlint`)
- remove benchmark shadowed `err` variables (`govet`) and clean pooled-reader error paths

+++ meta/isobmff/crx_test.go
@@ -41,7 +41,7 @@ func TestReadMetadataCR3XPacketCallback(t *testing.T) {
 	if !gotHeader.HasXMPMeta {
 		t.Fatal("expected HasXMPMeta=true")
 	}
-	if gotHeader.Length != uint32(len(xpacketPayload)) {
+	if gotHeader.Length != len(xpacketPayload) {
 		t.Fatalf("xpacket length = %d, want %d", gotHeader.Length, len(xpacketPayload))
 	}
 }
@@ -270,7 +270,7 @@ func TestReadMetadataHEIFXMPFromMdatCallback(t *testing.T) {
 	if !bytes.Equal(gotPayload, xmpPayload) {
 		t.Fatalf("xmp payload mismatch: got=%q want=%q", gotPayload, xmpPayload)
 	}
-	if gotHeader.Length != uint32(len(xmpPayload)) {
+	if gotHeader.Length != len(xmpPayload) {
 		t.Fatalf("xmp length = %d, want %d", gotHeader.Length, len(xmpPayload))
 	}
 	if !r.hasHave(metadataKindXMP) {
